Define `NavigatorManagedMedia` as a `partial interface`
The `NavigatorManagedMedia` is defined in Managed Configuration API. The Device
Attributes API can only extend it through a `partial interface` definition.

Note the addition of an entry to "link-defaults" is needed on a temporary
basis: due to the duplicate, the cross-references database currently has two
definitions of `NavigatorManagedMedia`, and Bikeshed would choose the wrong
one, creating a broken link. The entry can be removed once the spec has been
re-generated and the change propagated to the cross-references database (which
should take maximum 6 hours).